got rid of a lot of unnecessary code. shaved down updateVariables() as well. A lot of global variables that were unneeded are gone. Need to make sure appropriate C++ values get passed next

+++ src/cpp/Calculate_Magnetic_Moment_3D.cpp
@@ -393,7 +393,7 @@ JNIEXPORT void JNICALL Java_Calculate_1Magnetic_1Moment_13D_setMagMomentVariable
     return;
 }
 
-JNIEXPORT void JNICALL Java_Calculate_1Magnetic_1Moment_13D_setRi(JNIEnv *env, jobject thisObj, jint mri)
+JNIEXPORT void JNICALL Java_Calculate_1Magnetic_1Moment_13D_setRi(JNIEnv *env, jobject thisObj, jdouble mri)
 {
     m_Ri = mri;
     return;
